getGitHubStats function
